requirements.txt
fastapi
uvicorn
openai
jinja2
python-multipart
requests
python-dotenv
python-docx
reportlab
langgraph
